requirements.txt
Flask
tensorflow
Pillow
cassandra-driver
